requirements.txt
pandas
uuid
gnews
newspaper3k
nltk
psycopg2-binary
lxml[html_clean]
openpyxl
sqlalchemy
plotly
ipython

